+++ script.js
@@ -1,4 +1,4 @@
-import renderTables from "./tableRender.js"
-import buttonsFunctionality from "./buttonFunctionality.js"
+import renderTables from "./render/tableRender.js"
+import buttonsFunctionality from "./buttons/buttonFunctionality.js"
 renderTables()
 buttonsFunctionality()
